move test_quicly_set_cc to test.c since it's a small feature testing
