+++ t/test.h
@@ -60,6 +60,5 @@ void test_stream_concurrency(void);
 void test_received_cid(void);
 void test_local_cid(void);
 void test_retire_cid(void);
-void test_set_cc(void);
 
 #endif
